requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
idna==2.9
Jinja2==2.11.2
jinja2-markdown==0.0.3
libsass==0.20.0
mailchimp3==3.0.14
Markdown==3.2.1
MarkupSafe==1.1.1
python-frontmatter==0.5.0
python-slugify==4.0.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
text-unidecode==1.3
urllib3==1.25.9
validators==0.15.0
-e git+https://github.com/richardjpope/govuk-jinja-components@929a494a555d4f6abc20d3c22d2edeb62b7ae020#egg=govuk_jinja_components